pyproject.toml
[project]
name = "palantir"
version = "0.1.0"
description = "Automated data pipeline for Telegram content curation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.26.0",
    "aiosqlite>=0.22.1",
    "feedparser>=6.0.12",
    "google-genai>=1.68.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "telethon>=1.42.0",
]

[project.scripts]
palantir = "palantir.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/palantir"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.7",
]
